+++ src/ui/menu/settings/keybind_settings_popup.cpp
@@ -1,5 +1,6 @@
 #include "keybind_settings_popup.hpp"
 
+#include "keybind_setup_popup.hpp"
 #include <ui/general/list/list.hpp>
 #include <managers/settings.hpp>
 #include <util/ui.hpp>
@@ -8,18 +9,102 @@ using namespace geode::prelude;
 
 namespace {
     class KeybindCell : public CCLayer {
+        enumKeyCodes* settingPtr;
+        enumKeyCodes settingDefault;
+        CCMenu* menu;
+        CCNode* keyWrapper = nullptr;
+
     public:
         bool init(const char* name, const char* description, enumKeyCodes* ptr, enumKeyCodes defaultv) {
+            this->settingPtr = ptr;
+            this->settingDefault = defaultv;
+
+            float width = KeybindSettingsPopup::LIST_WIDTH;
+            float height = KeybindSettingsPopup::CELL_HEIGHT;
+
             Build<CCLabelBMFont>::create(name, "goldFont.fnt")
-                .pos(4.f, KeybindSettingsPopup::CELL_HEIGHT / 2.f)
+                .scale(0.6f)
+                .pos(6.f, height / 2.f)
                 .anchorPoint(0.f, 0.5f)
                 .parent(this);
 
-            this->setContentSize({KeybindSettingsPopup::LIST_WIDTH, KeybindSettingsPopup::CELL_HEIGHT});
+            Build<CCMenu>::create()
+                .layout(
+                    RowLayout::create()
+                        ->setAxisReverse(true)
+                        ->setAxisAlignment(AxisAlignment::End)
+                        ->setAutoScale(false)
+                )
+                .contentSize(width, height)
+                .anchorPoint(0.f, 0.f)
+                .pos(-6.f, 2.f)
+                .parent(this)
+                .store(menu);
+
+            auto spr = Build<CCSprite>::createSpriteName("pencil.png"_spr)
+                .collect();
+
+            Build<CircleButtonSprite>::create(spr)
+                .with([&](auto* item) {
+                    util::ui::rescaleToMatch(item, {height * 0.7f, height * 0.7f});
+                })
+                .intoMenuItem([this](auto) {
+                    KeybindSetupPopup::create(*this->settingPtr, [this](auto key) {
+                        *this->settingPtr = key;
+                        this->updateKey(key);
+                        GlobedSettings::get().save();
+                    })->show();
+                })
+                .parent(menu);
+
+            menu->updateLayout();
+
+            this->updateKey(*ptr);
+
+            this->setContentSize({width, height});
             return true;
         }
 
-        static KeybindCell* create(const char* name, const char* description, auto setting) {
+        void updateKey(enumKeyCodes key) {
+            auto& km = KeybindsManager::get();
+            auto formatted = globed::formatKey(km.convertCocosKey(key));
+
+            if (keyWrapper) {
+                keyWrapper->removeFromParent();
+            }
+
+            float sqsize = KeybindSettingsPopup::CELL_HEIGHT * 0.75f;
+            float sqscale = 2.f;
+
+            auto node = Build<CCNode>::create()
+                .parent(menu)
+                .store(keyWrapper)
+                .collect();
+
+            CCLabelBMFont* label = Build<CCLabelBMFont>::create(formatted, "goldFont.fnt")
+                .scale(0.63f)
+                .zOrder(2)
+                .parent(node)
+                .collect();
+
+            float width = label->getScaledContentWidth() + 6.f;
+
+            label->setPosition({width / 2.f + 1.f, sqsize / 2.f + 1.f});
+
+            auto bg = Build<CCScale9Sprite>::create("square02_small.png")
+                .contentSize(width * sqscale, sqsize * sqscale)
+                .scaleX(1.f / sqscale)
+                .scaleY(1.f / sqscale)
+                .opacity(80)
+                .pos(width / 2.f, sqsize / 2.f)
+                .parent(node)
+                .collect();
+
+            node->setContentSize(bg->getScaledContentSize());
+            menu->updateLayout();
+        }
+
+        static KeybindCell* create(const char* name, const char* description, auto& setting) {
             auto ret = new KeybindCell;
             if (ret->init(name, description, setting.ptr(), setting.Default)) {
                 ret->autorelease();
@@ -34,6 +119,8 @@ namespace {
 
 bool KeybindSettingsPopup::setup() {
     this->setTitle("Keybind Settings");
+
+    KeybindsManager::get().refreshBinds();
 
     auto rlayout = util::ui::getPopupLayoutAnchored(m_size);