+++ src/ui/menu/settings/keybind_settings_popup.hpp
@@ -8,7 +8,7 @@ public:
     static constexpr float POPUP_HEIGHT = 280.f;
     static constexpr float LIST_WIDTH = 340.f;
     static constexpr float LIST_HEIGHT = 200.f;
-    static constexpr float CELL_HEIGHT = 30.f;
+    static constexpr float CELL_HEIGHT = 26.f;
 
     static KeybindSettingsPopup* create();