+++ src/ui/menu/settings/keybind_setup_popup.cpp
@@ -5,29 +5,38 @@
 
 using namespace geode::prelude;
 
-bool KeybindSetupPopup::setup(void* settingStorage) {
+bool KeybindSetupPopup::setup(enumKeyCodes initialKey, Callback callback) {
     this->setTitle("Set Keybind");
+
+    this->originalKey = initialKey;
+    this->callback = std::move(callback);
 
     m_keybindLabel = CCLabelBMFont::create("Keybind: None", "bigFont.fnt");
     m_mainLayer->addChildAtPosition(m_keybindLabel, Anchor::Center);
 
     auto& gs = GlobedSettings::get();
 
-    int keybind = *(int*)(settingStorage);
-    this->settingStorage = (int*)settingStorage;
-
     auto applyButton = Build<ButtonSprite>::create("Apply", "bigFont.fnt", "GJ_button_01.png", 0.75f)
         .intoMenuItem([&gs, this] (auto) {
             if (!this->isValid) {
-                Notification::create("Please choose a valid key", NotificationIcon::Error, 0.5f)->show();
+                Notification::create("Please choose a valid key", NotificationIcon::Error, 0.75f)->show();
                 return;
             }
 
-            *this->settingStorage = this->key;
-            gs.save();
+            if (this->originalKey == this->key) {
+                this->removeFromParent();
+                return;
+            }
 
-            auto keystr = globed::formatKey((enumKeyCodes) this->key);
-            Notification::create(fmt::format("Set Keybind to {}", keystr), NotificationIcon::Success, 0.5f)->show();
+            auto& km = KeybindsManager::get();
+            if (km.isKeyUsed(this->key)) {
+                Notification::create("This key is already used by another action", NotificationIcon::Error, 1.0f)->show();
+                return;
+            }
+
+            if (this->callback) {
+                this->callback(this->key);
+            }
 
             this->removeFromParent();
         })
@@ -37,7 +46,7 @@ bool KeybindSetupPopup::setup(void* settingStorage) {
         .pos(0.f, 0.f)
         .parent(m_mainLayer);
 
-    this->keyDown((enumKeyCodes)*this->settingStorage);
+    this->keyDown(initialKey);
 
     return true;
 }
@@ -48,7 +57,12 @@ void KeybindSetupPopup::keyDown(enumKeyCodes keyCode) {
         return;
     }
 
-    this->key = (int)keyCode;
+    // for some reason sometimes this just randomly gets pressed every frame??
+    if (keyCode == -1) {
+        return;
+    }
+
+    this->key = keyCode;
 
     auto k = KeybindsManager::convertCocosKey(keyCode);
     this->isValid = k != globed::Key::None;
@@ -64,9 +78,9 @@ void KeybindSetupPopup::keyDown(enumKeyCodes keyCode) {
     m_keybindLabel->limitLabelWidth(POPUP_WIDTH - 16.f, 0.75f, 0.5f);
 }
 
-KeybindSetupPopup* KeybindSetupPopup::create(void* settingStorage) {
+KeybindSetupPopup* KeybindSetupPopup::create(enumKeyCodes initialKey, Callback callback) {
     auto ret = new KeybindSetupPopup();
-    if (ret->initAnchored(POPUP_WIDTH, POPUP_HEIGHT, settingStorage)) {
+    if (ret->initAnchored(POPUP_WIDTH, POPUP_HEIGHT, initialKey, std::move(callback))) {
         ret->autorelease();
         return ret;
     }