+++ src/ui/menu/settings/setting_cell.cpp
@@ -135,7 +135,6 @@ bool GlobedSettingCell::init(void* settingStorage, Type settingType, const char*
         case Type::PacketFragmentation: [[fallthrough]];
         case Type::AdvancedSettings: [[fallthrough]];
         case Type::LinkCode: [[fallthrough]];
-        case Type::Keybind: [[fallthrough]];
         case Type::ConnectionTest: [[fallthrough]];
         case Type::KeybindSettings: [[fallthrough]];
         case Type::AudioDevice: {
@@ -145,7 +144,6 @@ bool GlobedSettingCell::init(void* settingStorage, Type settingType, const char*
                 case Type::AdvancedSettings: text = "View"; break;
                 case Type::AudioDevice: text = "Set"; break;
                 case Type::LinkCode: text = "View"; break;
-                case Type::Keybind: text = "Set"; break;
                 case Type::ConnectionTest: text = "Test"; break;
                 case Type::KeybindSettings: text = "View"; break;
                 default: text = "what the figma"; break;
@@ -156,6 +154,7 @@ bool GlobedSettingCell::init(void* settingStorage, Type settingType, const char*
                 .intoMenuItem([this](auto* sender) {
                     this->onInteractiveButton(sender);
                 })
+                .scaleMult(1.1f)
                 .anchorPoint(0.5f, 0.5f)
                 .pos(CELL_WIDTH - 8.f, CELL_HEIGHT / 2)
                 .id("input-interactive-btn"_spr)
@@ -277,10 +276,7 @@ void GlobedSettingCell::onInteractiveButton(cocos2d::CCObject*) {
             }
             break;
         }
-        case Type::Keybind: {
-            KeybindSetupPopup::create(settingStorage)->show();
-            break;
-        }
+
         case Type::ConnectionTest: {
             ConnectionTestPopup::create()->show();
         } break;
@@ -432,7 +428,6 @@ void GlobedSettingCell::storeAndSave(std::any&& value) {
         case Type::PacketFragmentation: [[fallthrough]];
         case Type::InvitesFrom: [[fallthrough]];
         case Type::LinkCode: [[fallthrough]];
-        case Type::Keybind: [[fallthrough]];
         case Type::Int:
             *(int*)(settingStorage) = std::any_cast<int>(value); break;
         case Type::ConnectionTest: [[fallthrough]];